--- formal_language/test.py
@@ -1,2 +1,0 @@
-import formal
-
